Fixed the checkLargeAgeDifference in Family
There was an error in the checkLargeAgeDifference function in the Family
class. The calculation of the if statements and the ordering of the
Years.yearsBetween function was causing the wrong values to be
calculated and reported Jane Smith being twice as old as John Smith
which was incorrect. John Smith was actually twice as old as Jane Smith.